refactor: question_generator_v3.py에서 프롬프트 성격 코드 분리
